backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
openai>=1.3.5
cohere>=5.5.3
tiktoken>=0.5.2
markdown>=3.5.1
markdownify>=0.11.6
beautifulsoup4>=4.12.2
python-frontmatter>=1.0.0
redis>=5.0.1
slowapi>=0.1.8
pydantic-settings>=2.1.0
httpx>=0.24.0